Bundle JS/CSS dependencies for offline use
Inline all dependencies from template.html, including nbv.js which
previously had to be attached to the generator preview.

Note that the [inliner](https://github.com/remy/inliner) tool needs some
hacks to be called from a script and does not fail execution when a URL
could not be fetched. A more robust alternative would be great.

Fixes #5.